[ documents ] quite rough, but working..

+++ app/Http/Controllers/DocumentController.php
@@ -5,9 +5,14 @@ namespace App\Http\Controllers;
 use App\Document;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Input;
+use Illuminate\Http\Response;
 
 class DocumentController extends Controller
 {
+    private function checkUserOwnsDocument(Request $request, Document $document) {
+        return $request->user()->id == $document->user_id;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -24,26 +29,30 @@ class DocumentController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(Request $req)
     {
         $model = new Document();
-        if (!$model->validateAndFill($request->all())) {
+        if (!$model->validateAndFill($req->all())) {
             // Return failure code
-            return $errors;
+            return $model->errors();
         }
 
-        return $request->user()->documents()->save($model);
+        return $req->user()->documents()->save($model);
     }
 
     /**
      * Display the specified resource.
      *
+     * @param  \Illuminate\Http\Request  $req
      * @param  \App\Document  $document
      * @return \Illuminate\Http\Response
      */
-    public function show(Document $document)
+    public function show(Request $req, Document $document)
     {
-        // Does it belong to this user?
+        if (!$this->checkUserOwnsDocument($req, $document)) {
+            return response()->noContent(Response::HTTP_UNAUTHORIZED);
+        }
+
         return $document;
     }
 
@@ -51,22 +60,35 @@ class DocumentController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Document  $document
+     * @param  \App\Document  $documentument
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Document $document)
+    public function update(Request $req, Document $document)
     {
-        //
+        if (!$this->checkUserOwnsDocument($req, $document)) {
+            return response()->noContent(Response::HTTP_UNAUTHORIZED);
+        }
+
+        if (!$document->validateAndFill($req->all())) {
+            return $model->errors();
+        }
+
+        $document->save();
+        return $document;
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Document  $document
+     * @param  \App\Document  $documentument
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Document $document)
+    public function destroy(Request $req, Document $document)
     {
-        $document->delete();
+        if (!$this->checkUserOwnsDocument($req, $document)) {
+            return response()->noContent(Response::HTTP_UNAUTHORIZED);
+        }
+
+        return $document->delete();
     }
 }
